+++ pkg/ruler/ruler_test.go
@@ -260,7 +260,7 @@ func TestNotifierSendsUserIDHeader(t *testing.T) {
 	manager := newManager(t, cfg)
 	defer manager.Stop()
 
-	n, err := manager.getOrCreateNotifier("1")
+	n, err := manager.getOrCreateNotifier("1", manager.registry)
 	require.NoError(t, err)
 
 	// Loop until notifier discovery syncs up
@@ -275,10 +275,10 @@ func TestNotifierSendsUserIDHeader(t *testing.T) {
 
 	// Ensure we have metrics in the notifier.
 	assert.NoError(t, prom_testutil.GatherAndCompare(manager.registry.(*prometheus.Registry), strings.NewReader(`
-		# HELP cortex_prometheus_notifications_dropped_total Total number of alerts dropped due to errors when sending to Alertmanager.
-		# TYPE cortex_prometheus_notifications_dropped_total counter
-		cortex_prometheus_notifications_dropped_total{user="1"} 0
-	`), "cortex_prometheus_notifications_dropped_total"))
+		# HELP prometheus_notifications_dropped_total Total number of alerts dropped due to errors when sending to Alertmanager.
+		# TYPE prometheus_notifications_dropped_total counter
+		prometheus_notifications_dropped_total 0
+	`), "prometheus_notifications_dropped_total"))
 }
 
 func TestRuler_Rules(t *testing.T) {